+++ meta_package_manager/tests/test_cli_restore.py
@@ -17,6 +17,7 @@
 
 import pytest
 
+from ..managers import CURRENTLY_SUPPORTED_MANAGERS
 from .conftest import MANAGER_IDS
 from .test_cli import CLISubCommandTests
 
@@ -57,7 +58,7 @@ class TestRestore(CLISubCommandTests):
         assert "all-managers.toml" in result.stderr
         self.check_manager_selection(result)
 
-    @pytest.mark.parametrize("mid", MANAGER_IDS)
+    @pytest.mark.parametrize("mid", CURRENTLY_SUPPORTED_MANAGERS)
     def test_single_manager(self, invoke, create_toml, mid):
         toml_path = create_toml(
             "all-managers.toml",