+++ meta_package_manager/tests/test_pool.py
@@ -1,0 +1,49 @@
+# Copyright Kevin Deldycke <kevin@deldycke.com> and contributors.
+# All Rights Reserved.
+#
+# This program is Free Software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+import pytest
+
+from ..managers import CURRENTLY_SUPPORTED_MANAGERS, pool
+from .conftest import MANAGER_IDS
+
+""" Test the pool and its content. """
+
+
+def test_manager_count():
+    """Check all implemented package managers are accounted for."""
+    assert len(pool()) == 15
+    assert len(pool()) == len(MANAGER_IDS)
+    assert MANAGER_IDS == set(pool())
+
+
+def test_cached_pool():
+    assert pool() == pool()
+    assert pool() is pool()
+
+
+def test_sorted_pool():
+    assert list(pool()) == sorted([m.id for m in pool().values()])
+
+
+@pytest.mark.parametrize("manager_id", CURRENTLY_SUPPORTED_MANAGERS)
+def test_supported_managers(manager_id):
+    assert pool()[manager_id].supported is True
+
+
+@pytest.mark.parametrize("manager_id", MANAGER_IDS - CURRENTLY_SUPPORTED_MANAGERS)
+def test_unsupported_managers(manager_id):
+    assert pool()[manager_id].supported is False